Adds tests for action helpers ... need to fix tests for search results. Not sure how to deal with polling
